moves schedule hook and queries together
