+++ hooks/schedule/queries.ts
@@ -1,15 +1,15 @@
-import { ApiAllLocationsResponse as ApiLocationsResponse } from '../api/queries/locations/route'
-import { ApiUserFullProfileResponse } from '../api/queries/profiles/[userId]/route'
-import { ApiCurrentUserFullProfileResponse } from '../api/queries/profiles/current/route'
-import { ApiAllFullProfilesResponse } from '../api/queries/profiles/route'
-import { ApiUserRsvpsResponse } from '../api/queries/rsvps/[userId]/route'
-import { ApiCurrentUserRsvpsResponse } from '../api/queries/rsvps/current/route'
-import { ApiRsvpsResponse } from '../api/queries/rsvps/route'
-import { ApiUserSessionBookmarksResponse } from '../api/queries/session-bookmarks/[userId]/route'
-import { ApiCurrentUserSessionBookmarksResponse } from '../api/queries/session-bookmarks/current/route'
-import { ApiAllSessionBookmarksResponse } from '../api/queries/session-bookmarks/route'
-import { ApiSessionResponse } from '../api/queries/sessions/[id]/route'
-import { ApiAllSessionsResponse } from '../api/queries/sessions/route'
+import { ApiAllLocationsResponse as ApiLocationsResponse } from '../../app/api/queries/locations/route'
+import { ApiUserFullProfileResponse } from '../../app/api/queries/profiles/[userId]/route'
+import { ApiCurrentUserFullProfileResponse } from '../../app/api/queries/profiles/current/route'
+import { ApiAllFullProfilesResponse } from '../../app/api/queries/profiles/route'
+import { ApiUserRsvpsResponse } from '../../app/api/queries/rsvps/[userId]/route'
+import { ApiCurrentUserRsvpsResponse } from '../../app/api/queries/rsvps/current/route'
+import { ApiRsvpsResponse } from '../../app/api/queries/rsvps/route'
+import { ApiUserSessionBookmarksResponse } from '../../app/api/queries/session-bookmarks/[userId]/route'
+import { ApiCurrentUserSessionBookmarksResponse } from '../../app/api/queries/session-bookmarks/current/route'
+import { ApiAllSessionBookmarksResponse } from '../../app/api/queries/session-bookmarks/route'
+import { ApiSessionResponse } from '../../app/api/queries/sessions/[id]/route'
+import { ApiAllSessionsResponse } from '../../app/api/queries/sessions/route'
 
 // Utility function to handle API errors with detailed information
 const handleApiError = async (
